requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
minio==7.2.14
apache-airflow-providers-docker==4.0.0
apache-airflow-providers-slack==9.0.2